src/requirements.txt
Django
django-ninja
